api: add spec field to configure target group ipType
The field targetGroupIPType is added to the loadbalancer spec to allow
configuring ip address type of target group for API load balancers.

This field is not applicable to Classic Load Balancers (CLB).

This commit also defines a new network status field to determine the ip
type of API load balancers.